+++ src/main/java/org/cojen/tupl/Table.java
@@ -359,22 +359,21 @@ public interface Table<R> extends Closeable {
     /**
      * Stores the given row when a corresponding row doesn't exist.
      *
-     * @return false if a corresponding row already exists and nothing was inserted
      * @throws IllegalStateException if any required columns aren't set
-     * @throws UniqueConstraintException if a conflicting alternate key exists
+     * @throws UniqueConstraintException if a conflicting primary or alternate key exists
      */
-    public default boolean insert(Transaction txn, R row) throws IOException {
+    public default void insert(Transaction txn, R row) throws IOException {
         throw new UnmodifiableViewException();
     }
 
     /**
      * Stores the given row when a corresponding row already exists.
      *
-     * @return false if a corresponding row doesn't exist
      * @throws IllegalStateException if any required columns aren't set
+     * @throws NoSuchRowException if a corresponding row doesn't exist
      * @throws UniqueConstraintException if a conflicting alternate key exists
      */
-    public default boolean replace(Transaction txn, R row) throws IOException {
+    public default void replace(Transaction txn, R row) throws IOException {
         throw new UnmodifiableViewException();
     }
 
@@ -382,11 +381,11 @@ public interface Table<R> extends Closeable {
      * Updates an existing row with the modified columns of the given row, but the resulting
      * row isn't loaded back.
      *
-     * @return false if a corresponding row doesn't exist
      * @throws IllegalStateException if primary key isn't fully specified
+     * @throws NoSuchRowException if a corresponding row doesn't exist
      * @throws UniqueConstraintException if a conflicting alternate key exists
      */
-    public default boolean update(Transaction txn, R row) throws IOException {
+    public default void update(Transaction txn, R row) throws IOException {
         throw new UnmodifiableViewException();
     }
 
@@ -394,11 +393,11 @@ public interface Table<R> extends Closeable {
      * Updates an existing row with the modified columns of the given row, and then loads the
      * result back into the given row.
      *
-     * @return false if a corresponding row doesn't exist
      * @throws IllegalStateException if primary key isn't fully specified
+     * @throws NoSuchRowException if a corresponding row doesn't exist
      * @throws UniqueConstraintException if a conflicting alternate key exists
      */
-    public default boolean merge(Transaction txn, R row) throws IOException {
+    public default void merge(Transaction txn, R row) throws IOException {
         throw new UnmodifiableViewException();
     }